Completed self removal process
Added functionality to completely self-uninstall and delete service. A
viewer can connect to a machine, and send a signal to the machine to
uninstall and delete itself completely removing all traces of the
service.

+++ RemoteDesktop_Library/Timer.h
@@ -17,8 +17,8 @@ public:
 	explicit Timer(bool s = false){ if (s) Start(); }
 	void Start() { QueryPerformanceCounter(&m_start); }
 	void Stop() { QueryPerformanceCounter(&m_stop); }
-	long long Elapsed_micro() const{ return ((m_stop.QuadPart - m_start.QuadPart - m_overhead) * 1000000.0) / m_freq.QuadPart; }
-	long long Elapsed_milli() const{ return ((m_stop.QuadPart - m_start.QuadPart - m_overhead) * 1000.0) / m_freq.QuadPart; }
+	long long Elapsed_micro() const{ return (long long)((m_stop.QuadPart - m_start.QuadPart - m_overhead) * 1000000.0) / m_freq.QuadPart; }
+	long long Elapsed_milli() const{ return (long long)((m_stop.QuadPart - m_start.QuadPart - m_overhead) * 1000.0) / m_freq.QuadPart; }
 
 	template <typename T, typename Traits>
 	friend std::basic_ostream<T, Traits>& operator<<(std::basic_ostream<T, Traits>& out, const Timer& timer){ return out << timer.Elapsed().count(); }
